+++ deltafi-ui/src/mocks/graphql/getEnrichFlowByName.ts
@@ -55,6 +55,18 @@ const generateData = () => {
       version: "0.95.4",
     },
     flowStatus: flowStatusMap.get(flowStatus[Math.floor(Math.random() * flowStatus.length)]),
+    domainActions: [
+      {
+        name: "stix2_1.Stix2_1DomainAction",
+        type: "org.deltafi.stix.actions.Stix2_1DomainAction",
+        requiresDomains: ["stix"],
+        parameters: {
+          domains: {
+            stixDomain: "stix domain value",
+          },
+        },
+      },
+    ],
     enrichActions: [
       {
         name: "artificial-enrichment.BinaryEnrichAction",